Fix output file names, clean up more unused code
Removes a bunch more unused or commented-out
commands and environment variables that are
not relevant to the MyPaint appimage.

Refactor some slightly convoluted stuff and rename
some variables to be more descriptive.

The output appimage file is now  named with the
latest tag and a time stamp w. hours and minutes, so
that building multiple versions on the same day does not
require renaming.