Remove appimage_exec_wrapper
This component was never used in _this_ appimage, and with the recent
setup did not resolve the problems with starting external processes
(for external editing). This may have been an issue with the order of
initializations in AppRun, however, I did not try it out extensively
since the matter was solved in another way.