Patch mypaint to strip non-bundled-module paths
Since the build culls a lot of modules that are not really needed,
conflicts can arise if those culled modules are available locally.
Particularly local user-installs of those modules, since those
search paths were not shadowed by setting PYTHONPATH/PYTHONHOME.

Stripping out all paths that are not prefixed with the directory the
appimage is extracted to should solve it, unless the path is modified
by some other module prior to the import of the conflicting modules.